Remove image resource signature check
The code was looking for the Photoshop signature (8BIM).  Attempting to
load a PSD file created in PhotoDeluxe (signature PHUT) led to an
exception.

I do not really want to chase PSD issues in applications other than
Photoshop, especially one like PhotoDeluxe, which was discontinued in
2002.  However, there is no need to completely exclude non-Photoshop
PSDs from being loaded in.