Work item 7975: LayerGroup shown as layers
Layer groups are stored in PSD files in a backward-compatible manner, as
two empty layers surrounding the layers within the group.  The group
name is stored in the starting layer's name, and "</Layer Group>" is
stored in the ending layer's name.

Paint.NET does not have layer groups, so hierarchical layers in PSD
files must be loaded in as a flat list.  However, since layers are
stored in bottom-to-top order, the quasi-XML tag showed up first in a
flat list of layers.  The name of the group might be 100 layers down,
and the bracketing layers were indistinguishable from any other empty
layer.  This was quite confusing to users.

We can do better, by using the layer group's name in both cases, and
prefixing it with "Layer Group" or "End Layer Group."  This at least
gives the user some idea of the layer group hierarchy.

The text has been localized into all languages that Paint.NET is
localized into.  To preserve the single-DLL nature of the PSD plugin,
this is done by compiling in one resource file, and prefixing resource
strings with the language code.