Fix load error on double-byte systems
The length of Pascal strings in the PSD file is specified in bytes.
However, the code had been using this as the number of characters to
read.  If the string had double-byte characters, and Paint.NET was
running on a system with a double-byte Windows codepage, the code would
then read past the end of the string.

This has been fixed.  In addition, Pascal strings are now encoded using
the current Windows code page, so that the non-Unicode layer name is not
stored as question marks.