Fix a number of bugs in saving complex PSD files
When loading PSD files into Paint.NET, not all the features can
translate into the Paint.NET Document.  Similarly, when saving from
Paint.NET, a new PsdFile is created from scratch.  Thus, some of the
code paths were never exercised.  It was not possible to round-trip a
PSD file without losing data.  Small glitches in padding sometimes
caused the read position to be offset in Photoshop, so that it would
refuse to load the file.  Many of these glitches were the result of
incorrect or inadequate documentation in the PSD file spec.

I have used a large set of complex PSD files to thoroughly exercise the
code, and make sure that they can be loaded into a PsdFile and then
immediately saved to be bit-for-bit identical.  This allows the code to
be used to make simple modifications to the PSD file, such as renaming a
layer, without losing any data.

Note that bit-for-bit identity is only attained for 8-bit color depth,
since code has not been written to save higher-bitdepth images.  It also
requires that the Windows locale be the same as the one used to save the
original PSD file.

Changes:
  1. Fixed many issues with padding.
  2. More robust handling of the Additional Layer Info section.
  3. Stored data that was previously discarded, so that it can be passed
     through unchanged.  For example: thumbnails and RLE headers.
  4. Updated Image Resource IDs from Photoshop CS to CS6.
  5. Lots of other bugfixes and code cleanup.